Record (and serve) a human-readable build time
CO-512

This commit replaces the time property in build.properties and the info
resource with a human-readable ISO 8601 timestamp. The unixTime property
holds the old value, the number of milliseconds since the unix epoch.
I've kept it around because it is used in the filename of the jar, so is
nice to have handy if you're trying to figure out which jar is running.