Create Application class
CO-748

We've started to accumulate a lot of common startup logic in the
SkeletonApplication class. But since each application has its own
FooApplication class, and deletes the SkeletonApplication class,
changes in SkeletonApplication are not picked up when the skeleton
is merged back into each app. The changes have to be made manually,
which in practice means they are not made at all.

Introduce a new edu.oregonstate.mist.api.Application class to act
as a base class for all our applications. The setup method performs
all the common application startup logic. Each application should call
setup at the start of their run method