ignore page numbers and sizes equal to zero
CO-949

If a user gives a page number or size equal to zero, ignore it and use
the default page number or size.

Page size must not be zero because that can lead to division by zero
errors when we calculate pagination. (CO-949)

Page number must not be zero because we consider it to be 1-indexed, and
going below that doesn't make sense.